feat(statistics): 更新统计数据类型枚举值并调整相关引用

+++ only-danmuku-adapter/src/main/kotlin/edu/only4/danmuku/adapter/portal/api/payload/UCenterGetVideoCountInfo.kt
@@ -9,10 +9,10 @@ object UCenterGetVideoCountInfo {
 
     data class Response(
         /** 审核通过数 */
-        var auditPassCount: Int? = null,
+        var auditPassCount: Long? = null,
         /** 审核失败数 */
-        var auditFailCount: Int? = null,
+        var auditFailCount: Long? = null,
         /** 进行中数量 */
-        var inProgress: Int? = null
+        var inProgress: Long? = null,
     )
 }
